ai-deployment/container/sage-maker/yolo/code/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
pillow
numpy
opencv-python-headless==4.8.0.74
